% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{nbl_box_team}
\alias{nbl_box_team}
\title{Get team box score data from the NBL}
\usage{
nbl_box_team(season)
}
\arguments{
\item{season}{season(s) data wanted for, in the format '2021-2022'}
}
\value{
returns a dataframe of team box score data
}
\description{
Returns a data frame of all available team box scores
data for NBL matches played since 2015-16
}
\examples{
\dontrun{
try({
team_box_2022 <- nbl_box_team(season="2021-2022")
team_box_multiple <- nbl_box_team(season = c("2021-2022", "2022-2023"))
})
}
}
