% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{nbl_results}
\alias{nbl_results}
\title{Get match results for all NBL matches}
\usage{
nbl_results(wide_or_long)
}
\arguments{
\item{wide_or_long}{either 'wide' or 'long' to get either one row
per match, or a row for each team for each match}
}
\value{
returns a dataframe of match results
}
\description{
Returns a data frame of all results in either wide or long form
of all NBL matches played since 1979
}
\examples{
try({
results_wide <- nbl_results(wide_or_long="wide")
results_long <- nbl_results(wide_or_long="long")
})
}
