% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{nbl_box_player}
\alias{nbl_box_player}
\title{Get player box score data from the NBL}
\usage{
nbl_box_player()
}
\value{
returns a dataframe of player box score data
}
\description{
Returns a data frame of all available player box scores
data for NBL matches played since 2015-16
}
\examples{
try({
player_box <- nbl_box_player()
})
}
