% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{nbl_box_team}
\alias{nbl_box_team}
\title{Get team box score data from the NBL}
\usage{
nbl_box_team()
}
\value{
returns a dataframe of team box score data
}
\description{
Returns a data frame of all available team box scores
data for NBL matches played since 2015-16
}
\examples{
try({
team_box <- nbl_box_team()
})
}
