% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{nbl_pbp}
\alias{nbl_pbp}
\title{Get play-by-play data from the NBL}
\usage{
nbl_pbp()
}
\value{
returns a dataframe of play-by-play data
}
\description{
Returns a data frame of all available play-by-play
data for NBL matches played since 2015-16
}
\examples{
try({
pbp <- nbl_pbp()
})
}
